Moved bunch of files around.
