+++ slider/src/commands/addressCommand.h
@@ -2,7 +2,6 @@
 #define ADDRESSCOMMAND_H
 
 #include "src/core/menu/menuSystem.h"
-#include "src/app/brain/connectApp.h"
 #include <memory>
 
 using namespace Core;
