+++ slider/src/components/controllerConnector.cpp
@@ -1,4 +1,4 @@
-#include "connectApp.h"
+#include "controllerConnector.h"
 
 #include "src/debug.h"
 #include "src/core/network/address.h"
@@ -9,9 +9,7 @@
 using namespace Slider;
 using namespace Net;
 
-#define Led_Pin 17
 #define CONTROLLER_CONNECTION_DELAY 1000
-#define BRAIN_BROADCAST_DELAY 1000
 
 /*
                  ┌─────────┐                              ┌────────────┐
@@ -36,65 +34,6 @@ WAITING_FOR_HANDSHAKE  │                            Handshake   │  SENDING_H
 
 //const uint8_t receiver_mac[6] = { 0x94, 0x54, 0xc5, 0x63, 0x0a, 0xec };
 //const uint8_t sender_mac[6] = { 0x5c, 0x01, 0x3b, 0x68, 0xb1, 0x0c};
-
-BrainConnector::BrainConnector()
-    : state(ConnectionState::BROADCASTING),
-    isComplete(false)
-{}
-
-BrainConnector::~BrainConnector()
-{
-    for (const auto& handle : m_Callbacks)
-        WifiModule::GetInstance().RemoveReceiveCallback(handle);
-}
-
-void BrainConnector::Setup()
-{
-    LogInfo("Setup brain connector...");
-
-    auto connect = WifiModule::GetInstance().RegisterReceiveCallback<ConnectionRequest>("Brain-Connection",
-        [this](ConnectionRequest msg) {
-        if (state != ConnectionState::BROADCASTING) return;
-        LogInfo("Received: ", msg);
-        state = ConnectionState::SENDING_HANDSHAKE;
-        controllerMac = msg.from;
-    });
-    auto handshake = WifiModule::GetInstance().RegisterReceiveCallback<Handshake>("Brain-handshake",
-        [this](Handshake msg) {
-        if (state != ConnectionState::WAITING_FOR_HANDSHAKE) return;
-        LogInfo("Received: ", msg);
-        state = ConnectionState::CONNECTED;
-    });
-    m_Callbacks.push_back(connect);
-    m_Callbacks.push_back(handshake);
-    WifiModule::GetInstance().AddPeer(BROADCAST_ADDRESS);
-}
-
-void BrainConnector::Update()
-{
-    if (state == ConnectionState::BROADCASTING)
-    {
-        LogInfo("Sending connection request...");
-        ConnectionRequest request(WifiModule::GetInstance().GetMacAddress());
-        WifiModule::GetInstance().Send(BROADCAST_ADDRESS, request);
-        delay(BRAIN_BROADCAST_DELAY);
-    }
-    else if (state == ConnectionState::SENDING_HANDSHAKE)
-    {
-        WifiModule::GetInstance().RemovePeer(BROADCAST_ADDRESS);
-        WifiModule::GetInstance().AddPeer(controllerMac);
-        LogInfo("Sending handshake message...");
-        Handshake handshake(WifiModule::GetInstance().GetMacAddress());
-        WifiModule::GetInstance().Send(handshake);
-        state = ConnectionState::WAITING_FOR_HANDSHAKE;
-    }
-    else if (state == ConnectionState::CONNECTED)
-    {
-        //remove timer
-    }
-}
-
-//////////////////////////
 
 ControllerConnector::ControllerConnector() :
     state(ConnectionState::WAITING_FOR_CONNECTION),
